refactor: share product create shell
